Make it work on Windows.
Found by diffing usb traces from both nRF Connect and this lib. nRF Connect
was setting DTR and ppk2-rs wasn't. It seems Windows doesn't set it by default,
and the firmware refuses to reply if DTR is not set.

Clearing the buffers is not necessary in my limited testing, but it can't hurt,
so I'm leaving it in.